Signal Recorder and Display now work in double.
The numbers are too high for float.

+++ src/gui/src/SignalPlot.cpp
@@ -35,27 +35,28 @@ namespace ImGui
         {
             const ImU32 line_color = GetColorU32(ImGuiCol_PlotLines);
 
-            const float x_min = config.scale.x_scaled
-                                        ? config.scale.x_min
-                                        : config.values.x_series[config.values.offset];
-            const float x_max =
+            const double x_min = config.scale.x_scaled
+                                         ? config.scale.x_min
+                                         : config.values.x_series[config.values.offset];
+            const double x_max =
                     config.scale.x_scaled
                             ? config.scale.x_max
                             : config.values
                                       .x_series[config.values.offset + config.values.count - 1];
 
-            const float inverse_scale_x = (x_max == x_min) ? 0.f : (1.0f / (x_max - x_min));
-            const float inverse_scale_y =
+            const double inverse_scale_x = (x_max == x_min) ? 0.0 : (1.0 / (x_max - x_min));
+            const double inverse_scale_y =
                     (config.scale.y_min == config.scale.y_max)
-                            ? 0.0f
-                            : (1.0f / (config.scale.y_max - config.scale.y_min));
+                            ? 0.0
+                            : (1.0 / (config.scale.y_max - config.scale.y_min));
 
-            const float first_x_value = config.values.x_series[config.values.offset];
-            const float first_y_value = config.values.y_series[config.values.offset];
+            const double first_x_value = config.values.x_series[config.values.offset];
+            const double first_y_value = config.values.y_series[config.values.offset];
 
             ImVec2 first_normalized_point = ImVec2(
-                    ImSaturate((first_x_value - x_min) * inverse_scale_x),
-                    1.0f - ImSaturate((first_y_value - config.scale.y_min) * inverse_scale_y));
+                    ImSaturate(static_cast<float>((first_x_value - x_min) * inverse_scale_x)),
+                    1.0f - ImSaturate(static_cast<float>((first_y_value - config.scale.y_min) *
+                                                         inverse_scale_y)));
 
             if (first_x_value > x_min)
             {
@@ -73,11 +74,12 @@ namespace ImGui
             const int end_index = config.values.offset + config.values.count;
             for (int value_index = config.values.offset; value_index < end_index; value_index++)
             {
-                const float x_value = config.values.x_series[value_index] - x_min;
-                const float y_value = config.values.y_series[value_index];
+                const double x_value = config.values.x_series[value_index] - x_min;
+                const double y_value = config.values.y_series[value_index];
                 const ImVec2 second_normalized_point =
-                        ImVec2(ImSaturate(x_value * inverse_scale_x),
-                               1.0f - ImSaturate((y_value - config.scale.y_min) * inverse_scale_y));
+                        ImVec2(ImSaturate(static_cast<float>(x_value * inverse_scale_x)),
+                               1.0f - ImSaturate(static_cast<float>((y_value - config.scale.y_min) *
+                                                                    inverse_scale_y)));
 
                 ImVec2 first_position = ImLerp(inner_bounding_box.Min, inner_bounding_box.Max,
                                                first_normalized_point);
@@ -95,7 +97,6 @@ namespace ImGui
                 last_position.x = inner_bounding_box.Max.x;
                 window->DrawList->AddLine(second_position, last_position, line_color,
                                           config.line_thickness);
-
             }
         }
     }
