+++ src/gui/src/lib/SignalRecorder.h
@@ -13,14 +13,14 @@ public:
     explicit SignalRecorder(std::size_t size);
 
     [[nodiscard]] std::size_t size() const;
-    [[nodiscard]] const float* time_series() const;
-    [[nodiscard]] const float* state_series() const;
+    [[nodiscard]] const double* time_series() const;
+    [[nodiscard]] const double* state_series() const;
 
     void add(Edge edge);
 
 private:
-    std::vector<float> time_values;
-    std::vector<float> state_values;
+    std::vector<double> time_values;
+    std::vector<double> state_values;
 };
 
 #endif //MICRALN_SIGNALRECORDER_H
